Changed input of get roles of user and renamed it to better fit its name
